requirements.txt
pyzmq == 22.1.0
torch == 1.8.1+cu111
torchvision == 0.9.1+cu111
